Refactored using golang composition and interfaces.
